Move keymap functions to a separate file
